rosen_backend/requirements.txt
asgiref==3.8.1
bcrypt==4.1.2
Django==5.0.4
djangorestframework==3.15.1
jedi==0.19.1
parso==0.8.4
pillow==10.3.0
sqlparse==0.5.0
